Refactor CompletionHelper
Fix caching of PatternMatchers
Deleted CompletionHelperService
