+++ src/EditorFeatures/Core/IntelliSense/AsyncCompletion/ItemManager.CompletionListUpdater.cs
@@ -131,17 +131,19 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.IntelliSense.AsyncComplet
 
                 try
                 {
+                    using var patternMatchHelper = new PatternMatchHelper(_filterText);
+
                     // Determine the list of items to be included in the completion list.
                     // This is computed based on the filter text as well as the current
                     // selection of filters and expander.
-                    AddCompletionItems(itemsToBeIncluded, cancellationToken);
+                    AddCompletionItems(itemsToBeIncluded, patternMatchHelper, cancellationToken);
 
                     // Decide if we want to dismiss an empty completion list based on CompletionRules and filter usage.
                     if (itemsToBeIncluded.Count == 0)
                         return HandleAllItemsFilteredOut();
 
                     var highlightAndFilterTask = Task.Run(
-                        () => GetHighlightedListAndUpdatedFilters(session, itemsToBeIncluded, cancellationTokenSource.Token),
+                        () => GetHighlightedListAndUpdatedFilters(session, itemsToBeIncluded, patternMatchHelper, cancellationTokenSource.Token),
                         cancellationTokenSource.Token);
 
                     // Decide the item to be selected for this completion session.
@@ -181,9 +183,9 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.IntelliSense.AsyncComplet
                 }
 
                 (CompletionList<CompletionItemWithHighlight>, ImmutableArray<CompletionFilterWithState>) GetHighlightedListAndUpdatedFilters(
-                    IAsyncCompletionSession session, IReadOnlyList<MatchResult> itemsToBeIncluded, CancellationToken cancellationToken)
+                    IAsyncCompletionSession session, IReadOnlyList<MatchResult> itemsToBeIncluded, PatternMatchHelper patternMatchers, CancellationToken cancellationToken)
                 {
-                    var highLightedList = GetHighlightedList(session, itemsToBeIncluded, cancellationToken);
+                    var highLightedList = GetHighlightedList(session, patternMatchers, itemsToBeIncluded, cancellationToken);
                     var updatedFilters = GetUpdatedFilters(itemsToBeIncluded, cancellationToken);
                     return (highLightedList, updatedFilters);
                 }
@@ -228,7 +230,7 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.IntelliSense.AsyncComplet
                 }
             }
 
-            private void AddCompletionItems(List<MatchResult> list, CancellationToken cancellationToken)
+            private void AddCompletionItems(List<MatchResult> list, PatternMatchHelper patternMatchers, CancellationToken cancellationToken)
             {
                 // Convert initial and update trigger reasons to corresponding Roslyn type so 
                 // we can interact with Roslyn's completion system
@@ -260,9 +262,8 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.IntelliSense.AsyncComplet
                     // currentIndex is used to track the index of the VS CompletionItem in the initial sorted list to maintain a map from Roslyn item to VS item.
                     // It's also used to sort the items by pattern matching results while preserving the original alphabetical order for items with
                     // same pattern match score since `List<T>.Sort` isn't stable.
-                    if (CompletionHelper.TryCreateMatchResult(_completionHelper, itemData.RoslynItem, _filterText,
-                        roslynInitialTriggerKind, roslynFilterReason, _recentItemsManager.GetRecentItemIndex(itemData.RoslynItem), _highlightMatchingPortions, currentIndex,
-                        out var matchResult))
+                    if (patternMatchers.TryCreateMatchResult(itemData.RoslynItem, roslynInitialTriggerKind, roslynFilterReason,
+                        _recentItemsManager.GetRecentItemIndex(itemData.RoslynItem), _highlightMatchingPortions, currentIndex, out var matchResult))
                     {
                         list.Add(matchResult);
 
@@ -568,6 +569,7 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.IntelliSense.AsyncComplet
 
             private CompletionList<CompletionItemWithHighlight> GetHighlightedList(
                 IAsyncCompletionSession session,
+                PatternMatchHelper patternMatchers,
                 IReadOnlyList<MatchResult> matchResults,
                 CancellationToken cancellationToken)
             {
@@ -575,16 +577,13 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.IntelliSense.AsyncComplet
                 {
                     var vsItem = GetCorrespondingVsCompletionItem(matchResult, cancellationToken);
                     var highlightedSpans = _highlightMatchingPortions
-                        ? GetHighlightedSpans(matchResult, _completionHelper, _filterText)
+                        ? GetHighlightedSpans(matchResult, patternMatchers)
                         : ImmutableArray<Span>.Empty;
 
                     return new CompletionItemWithHighlight(vsItem, highlightedSpans);
                 }));
 
-                static ImmutableArray<Span> GetHighlightedSpans(
-                    MatchResult matchResult,
-                    CompletionHelper completionHelper,
-                    string filterText)
+                static ImmutableArray<Span> GetHighlightedSpans(MatchResult matchResult, PatternMatchHelper patternMatchers)
                 {
                     if (matchResult.CompletionItem.HasDifferentFilterText || matchResult.CompletionItem.HasAdditionalFilterTexts)
                     {
@@ -593,8 +592,8 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.IntelliSense.AsyncComplet
                         // However, if the Roslyn item's FilterText is different from its DisplayText, we need to do the match against the
                         // display text of the VS item directly to get the highlighted spans. This is done in a best effort fashion and there
                         // is no guarantee a proper match would be found for highlighting.
-                        return completionHelper.GetHighlightedSpans(
-                            matchResult.CompletionItem, filterText, CultureInfo.CurrentCulture).SelectAsArray(s => s.ToSpan());
+                        return patternMatchers.GetHighlightedSpans(matchResult.CompletionItem.GetEntireDisplayText(), CultureInfo.CurrentCulture)
+                            .SelectAsArray(s => s.ToSpan());
                     }
 
                     var patternMatch = matchResult.PatternMatch;
