--- src/Features/Core/Portable/Completion/ICompletionHelperService.cs
@@ -1,13 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using Microsoft.CodeAnalysis.Host;
-
-namespace Microsoft.CodeAnalysis.Completion
-{
-    internal interface ICompletionHelperService : IWorkspaceService
-    {
-        CompletionHelper GetCompletionHelper(Document document);
-    }
-}
